+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/internal/storage/db/H2DatabaseConfiguration.java
@@ -50,5 +50,10 @@ public class H2DatabaseConfiguration extends DatabaseConfiguration {
     public boolean hasUpdateReturning() {
       return false;
     }
+
+    @Override
+    public String castToEnumType(String variable, String type) {
+      return variable;
+    }
   }
 }