+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/internal/storage/db/MysqlDatabaseConfiguration.java
@@ -68,5 +68,10 @@ public class MysqlDatabaseConfiguration extends DatabaseConfiguration {
     public boolean hasUpdateReturning() {
       return false;
     }
+
+    @Override
+    public String castToEnumType(String variable, String type) {
+      return variable;
+    }
   }
 }