+++ nflow-engine/src/main/resources/scripts/db/postgresql.create.ddl.sql
@@ -27,10 +27,12 @@ create trigger update_nflow_modified before update on nflow_workflow for each ro
 drop index nflow_workflow_activation;
 create index nflow_workflow_activation on nflow_workflow(next_activation) where next_activation is not null;
 
+create type action_type as enum ('stateExecution', 'stateExecutionFailed', 'recovery', 'externalChange');
 create table if not exists nflow_workflow_action (
   id serial primary key,
   workflow_id int not null,
   executor_id int not null default -1,
+  type action_type not null default 'stateExecution',
   state varchar(64) not null,
   state_text varchar(128),
   retry_no int not null,