+++ nflow-engine/src/main/resources/scripts/db/update-1.2.0-x/h2.update.ddl.sql
@@ -1,0 +1,1 @@
+alter table nflow_workflow_action add type varchar(32) not null default 'stateExecution' check type in ('stateExecution', 'stateExecutionFailed', 'recovery', 'manualStateChange');