+++ nflow-engine/src/main/resources/scripts/db/update-1.2.0-x/mysql.update.ddl.sql
@@ -1,0 +1,1 @@
+alter table nflow_workflow_action add type enum('stateExecution', 'stateExecutionFailed', 'recovery', 'manualStateChange') not null default 'stateExecution';