+++ nflow-engine/src/main/resources/scripts/db/update-1.2.0-x/postgresql.update.ddl.sql
@@ -1,0 +1,2 @@
+create type action_type as enum ('stateExecution', 'stateExecutionFailed', 'recovery', 'manualStateChange');
+alter table nflow_workflow_action add type action_type not null default 'stateExecution';